requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
httpx==0.28.1

# Daten
pandas==2.3.1
pandas-ta-openbb==0.4.22

# Config + Cache + DB
pydantic-settings==2.2.1
fastapi-cache2[redis]==0.2.1
redis>=4.2.0,<5.0.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9        # psycopg für Postgres
